// apps/api/src/database/db.generated.ts
export type Generated<T> = T extends ColumnType<infer S, infer I, infer U>
  ? ColumnType<S, I | undefined, U>
  : ColumnType<T, T | undefined, T>;

export type Numeric = ColumnType<string, number | string, number | string>;

export type Timestamp = ColumnType<Date, Date | string, Date | string>;
// ...
export interface FinanceBudgetPlanVersions {
  createdAt: Generated<Timestamp>;
  deletedAt: Timestamp | null;
  effectiveFromMonth: string;
  id: Generated<string>;
  note: string | null;
  updatedAt: Generated<Timestamp>;
}